move files and fix multiline commands
