+++ src/cli/cli_params.py
@@ -1,5 +1,9 @@
 from enum import Enum
 from typing import Annotated
+from dataclasses import dataclass
+import questionary
+import os
+import sys
 
 from typer import Argument, Option
 
@@ -102,3 +106,376 @@ RobotName = Annotated[
 AnsiblePlaybook = Annotated[
     str, Option("-ap", "--ansible_playbook", help="Ansible playbook name")
 ]
+
+
+@dataclass
+class BuildConfig:
+    action: ActionArgument
+    search_query: str | None = None
+    no_optimized_build: bool = False
+    debug_build: bool = False
+    select_debug_binaries: list | None = None
+    flash_robots: list | None = None
+    ssh_password: str | None = None
+    interactive_search: bool = False
+    tracy: bool = False
+    test_suite: bool = False
+    enable_thunderscope: bool = False
+    stop_ai_on_start: bool = False
+    jobs_option: str | None = None
+    runs: int | None = None
+    robot_name: str | None = None
+    ansible_playbook: str | None = None
+    debug_powerloop: bool = False
+    disable_power_service: bool = False
+    disable_motor_service: bool = False
+
+
+class BazelFlag(tuple, Enum):
+    DEBUG_BUILD = ("-c", "dbg")
+    OPTIMIZED = ("--copt=-O3",)
+    ROBOT_PLATFORM = ("--platforms=//toolchains/cc:robot",)
+    TRACY = ("--cxxopt=-DTRACY_ENABLE",)
+    THUNDERSCOPE = ("--spawn_strategy=local", "--test_env=DISPLAY=:0")
+    NO_CACHE_TESTS = ("--cache_test_results=false",)
+    DEBUG_POWERLOOP = ("--//software/power:debug_powerloop",)
+    DISABLE_POWER_SERVICE = ("--//software/embedded:disable_power_service",)
+    DISABLE_MOTOR_SERVICE = ("--//software/embedded:disable_motor_service",)
+
+
+class InteractiveCli:
+    HISTORY_FILE = "/tmp/tbots_history"
+    HISTORY_MAX_ENTRIES = 50
+
+    # ---------------------------------------------------------------------------
+    # Interactive CLI styling
+    #
+    # Shared questionary style applied to every interactive prompt.
+    # ---------------------------------------------------------------------------
+    INTERACTIVE_STYLE = questionary.Style(
+        [
+            ("qmark", "fg:ansicyan bold"),
+            ("question", "bold"),
+            ("pointer", "fg:ansicyan bold"),
+            ("highlighted", "fg:ansicyan bold"),
+            ("selected", "fg:ansigreen"),
+            ("answer", "fg:ansicyan bold"),
+            ("text", "fg:ansibrightblack"),
+            ("instruction", "fg:ansibrightblack italic"),
+            ("disabled", "fg:ansibrightblack italic"),
+        ]
+    )
+
+    # ---------------------------------------------------------------------------
+    # Interactive CLI choices
+    #
+    # Each questionary.Choice pairs an option's display title with a description.
+    # questionary renders the description inline when the option is highlighted in
+    # the interactive menus, giving users guidance without leaving the prompt.
+    # When a Choice has no explicit value, questionary returns its title, so the
+    # titles below double as the values consumed by start_interactive_cli.
+    # ---------------------------------------------------------------------------
+
+    # Top-level "What would you like to do?" menu.
+    class Category(str, Enum):
+        THUNDERSCOPE = "thunderscope"
+        TEST = "test"
+        FLASH = "flash"
+        REPEAT_CMD_MSG = "Repeat a past command"
+
+    CATEGORY_CHOICES = [
+        questionary.Choice(
+            title="Run thunderscope",
+            value=Category.THUNDERSCOPE,
+            description="Launch Thunderscope against the simulator or real robots",
+        ),
+        questionary.Choice(
+            title="Test",
+            value=Category.TEST,
+            description="Run a single test or the entire test suite",
+        ),
+        questionary.Choice(
+            title="Flash",
+            value=Category.FLASH,
+            description="Deploy software or firmware to a robot via Ansible",
+        ),
+    ]
+
+    class LaunchMode(str, Enum):
+        SIM = "sim"  # This value is not used, only the enum is checked
+        RUN_DIAG = "run_diagnostics"
+        RUN_BLUE = "run_blue"
+        RUN_YELLOW = "run_yellow"
+
+    # Thunderscope "Launch mode?" menu.
+    LAUNCH_MODE_CHOICES = [
+        questionary.Choice(
+            title="Simulator",
+            value=LaunchMode.SIM,
+            description="Run Thunderscope against the simulated full system",
+        ),
+        questionary.Choice(
+            title="Diagnostics",
+            value=LaunchMode.RUN_DIAG,
+            description="Run Thunderscope in diagnostics mode against real robots",
+        ),
+        questionary.Choice(
+            title="Run blue",
+            value=LaunchMode.RUN_BLUE,
+            description="run as blue team",
+        ),
+        questionary.Choice(
+            title="Run yellow",
+            value=LaunchMode.RUN_YELLOW,
+            description="run as yellow team",
+        ),
+    ]
+
+    class SimOptions(str, Enum):
+        RECORD_STATS = "record_stats"
+        ENABLE_AUTOREF = "enable_autoref"
+        CI_MODE = "ci_mode"
+        ENABLE_REALISM = "enable_realism"
+        ENABLE_AUTOGC = "enable_autogc"
+
+    # Thunderscope simulator "Options:" checkbox.
+    THUNDERSCOPE_SIMULATOR_OPTION_CHOICES = [
+        questionary.Choice(
+            title="Enable Automatic Referee",
+            value=SimOptions.ENABLE_AUTOREF,
+            description="Run the autoref alongside the simulator",
+        ),
+        questionary.Choice(
+            title="CI Mode",
+            value=SimOptions.CI_MODE,
+            description="Run in continuous integration mode (headless friendly)",
+        ),
+        questionary.Choice(
+            title="Record Statistics",
+            value=SimOptions.RECORD_STATS,
+            description="Record gameplay statistics for a given duration in minutes",
+        ),
+        questionary.Choice(
+            title="Enable Realism",
+            value=SimOptions.ENABLE_REALISM,
+            description="Enable realistic simulation physics and sensor noise",
+        ),
+        questionary.Choice(
+            title="Enable Automatic Game Controller",
+            value=SimOptions.ENABLE_AUTOGC,
+            description="Run the automated game controller alongside the simulator",
+        ),
+    ]
+
+    # Deploy robot software "Options:" checkbox. Each option compiles Thunderloop
+    # with a preprocessor flag that disables the corresponding service, letting
+    # Thunderloop run on a robot whose powerboard or motorboard is unavailable.
+    class DeployOption(str, Enum):
+        DISABLE_POWER_SERVICE = "DISABLE_POWER_SERVICE"
+        DISABLE_MOTOR_SERVICE = "DISABLE_MOTOR_SERVICE"
+
+    DEPLOY_ROBOT_SOFTWARE_OPTION_CHOICES = [
+        questionary.Choice(
+            title="Disable Power Service",
+            value=DeployOption.DISABLE_POWER_SERVICE,
+            description="Compile Thunderloop without the Power Service (no powerboard)",
+        ),
+        questionary.Choice(
+            title="Disable Motor Service",
+            value=DeployOption.DISABLE_MOTOR_SERVICE,
+            description="Compile Thunderloop without the Motor Service (no motorboard)",
+        ),
+    ]
+
+    # Marker value returned by the DEBUG_POWERLOOP playbook choice. It maps to the
+    # deploy_powerboard.yml playbook but additionally compiles powerloop_main with
+    # the DEBUG_POWERLOOP flag, swapping in bare setup()/loop() stubs so arbitrary
+    # code can be flashed onto the powerboard microcontroller for debugging.
+    DEBUG_POWERLOOP_PLAYBOOK = "deploy_powerboard.yml (DEBUG_POWERLOOP)"
+
+    # Flash "Select playbook:" menu.
+    PLAYBOOK_CHOICES = [
+        questionary.Choice(
+            title="setup_pi.yml",
+            value=("setup_pi.yml", False),
+            description="First-time setup of the Raspberry Pi on a robot",
+        ),
+        # If the playbook name changes then you must change the if statement below.
+        questionary.Choice(
+            title="deploy_robot_software.yml",
+            value=("deploy_robot_software.yml", False),
+            description="Build and flash Thunderloop and the robot software",
+        ),
+        questionary.Choice(
+            title="deploy_powerboard.yml",
+            value=("deploy_powerboard.yml", False),
+            description="Flash the powerboard firmware (powerloop_main)",
+        ),
+        questionary.Choice(
+            title=DEBUG_POWERLOOP_PLAYBOOK,
+            value=("deploy_powerboard.yml", True),
+            description="Flash powerloop_main built with the DEBUG_POWERLOOP flag "
+            "for inserting arbitrary debug code onto the powerboard",
+        ),
+        questionary.Choice(
+            title="deploy_motor_firmware.yml",
+            value=("deploy_motor_firmware.yml", False),
+            description="Flash the motor controller firmware",
+        ),
+    ]
+
+    @staticmethod
+    def load_history() -> list[str]:
+        if not os.path.exists(InteractiveCli.HISTORY_FILE):
+            return []
+        with open(InteractiveCli.HISTORY_FILE) as f:
+            lines = [line.strip() for line in f.readlines()]
+        return [l.replace("\\n", "\n") for l in lines if l]
+
+    @staticmethod
+    def save_to_history(cmd_title: str, cmd_str: str):
+        history = InteractiveCli.load_history()
+        history = [h for h in history if h != cmd_str]
+        history.append(cmd_title + ": " + cmd_str)
+        history = history[-InteractiveCli.HISTORY_MAX_ENTRIES :]
+        with open(InteractiveCli.HISTORY_FILE, "w") as f:
+            f.write("\n".join(h.replace("\n", "\\n") for h in history) + "\n")
+
+    @staticmethod
+    def start_interactive_cli() -> tuple[str, BuildConfig, list[str]] | None:
+        """Run the menu-driven interactive CLI.
+
+        Walks the user through a series of questionary prompts to assemble a
+        :class:`BuildConfig`, then validates, builds, and executes the resulting
+        Bazel command. Returns early without running anything if the user aborts
+        the top-level prompt.
+
+        :returns The build title, build config assembled from the user prompts, and
+        extra args. None if failure to build config.
+        """
+        config = BuildConfig(action=ActionArgument.run)  # Default action
+        extra_args = []
+
+        history = InteractiveCli.load_history()
+        choices = InteractiveCli.CATEGORY_CHOICES
+        if history:
+            choices = [
+                InteractiveCli.Category.REPEAT_CMD_MSG.value
+            ] + InteractiveCli.CATEGORY_CHOICES
+
+        category = questionary.select(
+            "What would you like to do?",
+            choices=choices,
+            style=InteractiveCli.INTERACTIVE_STYLE,
+        ).unsafe_ask()
+
+        cmd_title = ""
+
+        match category:
+            case InteractiveCli.Category.REPEAT_CMD_MSG:
+                choices = [
+                    questionary.Choice(
+                        title=entry.split(": ")[0],
+                        value=entry.split(": ")[1],
+                        description=entry.split(": ")[1],
+                    )
+                    for entry in list(reversed(history))
+                ]
+                past_cmd = questionary.select(
+                    "Select a command to re-run:",
+                    choices=choices,
+                ).unsafe_ask()
+
+                if not past_cmd:
+                    return
+                print(f"\n{'=' * 33} Running: {'=' * 38}\n\n{past_cmd}\n\n{'=' * 81}\n")
+                code = os.system(past_cmd)
+                sys.exit(1 if code != 0 else 0)
+
+            case InteractiveCli.Category.THUNDERSCOPE:
+                config.action = ActionArgument.run
+                config.search_query = "thunderscope"
+                cmd_title = "Run Thunderscope"
+                launch = questionary.select(
+                    "Launch mode?",
+                    choices=InteractiveCli.LAUNCH_MODE_CHOICES,
+                    style=InteractiveCli.INTERACTIVE_STYLE,
+                ).unsafe_ask()
+                if launch == InteractiveCli.LaunchMode.SIM:
+                    selected = questionary.checkbox(
+                        "Options:",
+                        choices=InteractiveCli.THUNDERSCOPE_SIMULATOR_OPTION_CHOICES,
+                        style=InteractiveCli.INTERACTIVE_STYLE,
+                    ).unsafe_ask()
+                    for opt in selected:
+                        extra_args.extend([f"--{opt.value}"])
+                        if opt == InteractiveCli.SimOptions.RECORD_STATS:
+                            duration = questionary.text(
+                                "Enter record stats duration (minutes):",
+                                style=InteractiveCli.INTERACTIVE_STYLE,
+                                validate=lambda x: x.isdigit(),
+                            ).unsafe_ask()
+                            extra_args.extend([duration])
+                else:
+                    iface = questionary.text(
+                        "Network interface?", style=InteractiveCli.INTERACTIVE_STYLE
+                    ).unsafe_ask()
+                    extra_args.extend([f"--{launch.value}", "--interface", iface])
+
+            case InteractiveCli.Category.TEST:
+                cmd_title = "Run Tests"
+                config.action = ActionArgument.test
+                test_name = questionary.text(
+                    "Enter test name (leave empty for entire suite)",
+                    style=InteractiveCli.INTERACTIVE_STYLE,
+                ).unsafe_ask()
+                if not test_name:
+                    config.test_suite = True
+                else:
+                    config.search_query = test_name
+                    runs_str = questionary.text(
+                        "Number of times to run each test (leave empty for 1):",
+                        style=InteractiveCli.INTERACTIVE_STYLE,
+                    ).unsafe_ask()
+                    if runs_str and runs_str.isdigit() and int(runs_str) > 1:
+                        config.runs = int(runs_str)
+
+            case InteractiveCli.Category.FLASH:
+                cmd_title = "Flash"
+                config.action = ActionArgument.run
+                config.search_query = "ansible"
+                playbook, debug_powerloop = questionary.select(
+                    "Select playbook:",
+                    choices=InteractiveCli.PLAYBOOK_CHOICES,
+                    style=InteractiveCli.INTERACTIVE_STYLE,
+                ).unsafe_ask()
+                cmd_title += " " + playbook
+                config.ansible_playbook = playbook
+                config.debug_powerloop = debug_powerloop
+
+                if config.ansible_playbook == "deploy_robot_software.yml":
+                    selected = (
+                        questionary.checkbox(
+                            "Options:",
+                            choices=InteractiveCli.DEPLOY_ROBOT_SOFTWARE_OPTION_CHOICES,
+                            style=InteractiveCli.INTERACTIVE_STYLE,
+                        ).unsafe_ask()
+                        or []
+                    )
+                    config.disable_power_service = (
+                        InteractiveCli.DeployOption.DISABLE_POWER_SERVICE in selected
+                    )
+                    config.disable_motor_service = (
+                        InteractiveCli.DeployOption.DISABLE_MOTOR_SERVICE in selected
+                    )
+
+                name = questionary.text(
+                    "Robot name?", style=InteractiveCli.INTERACTIVE_STYLE
+                ).unsafe_ask()
+                config.robot_name = name
+                cmd_title += " to " + name
+                config.ssh_password = questionary.password(
+                    "SSH password?", style=InteractiveCli.INTERACTIVE_STYLE
+                ).unsafe_ask()
+
+        return cmd_title, config, extra_args
